requirements.txt
coreapi==1.32.0
Django==1.9.2
django-cors-headers==1.1.0
django-logging-json==1.5.3
django-rest-swagger==2.0.2
djangorestframework==3.3.2
drfdocs==0.0.11
elasticsearch==2.3.0
itypes==1.1.0
Markdown==2.6.6
openapi-codec==1.0.0
requests==2.10.0
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
urllib3==1.16
wheel==0.24.0
